feat(electron): relocate pin to global menu
